Create GenericAllClassifiers.py
This code runs all the generic classifiers on the data, giving you an idea of how the models are performing. This is without parameter tuning and can be used as a starting step to model building and using this maybe we can choose one model and build on it further.