Add files via upload
This file will be used to identify the number of neurons and layers for an ANN along with the learning rate. This is replicated from Jeff Heaton's course on deep learning. His code was for multi-class classification and this one is for binary classification so there are a few changes here and there. The code overall is an inspiration and savior.